Datatype (#363)
* Add support for changing dataytpe in DeepAREstimator

* Make dtype terminology consistent

* undo unnecessary changes added for supporting dtype

* need new mxnet version for backward to succeed

* fix float_type in backtest

* pass dtype explicitly to sample function

* Update smoke-test to include tests for dtype

+++ src/gluonts/distribution/transformed_distribution.py
@@ -17,6 +17,7 @@ from typing import Optional, Tuple, List
 # Third-party imports
 from mxnet import autograd
 import mxnet as mx
+import numpy as np
 
 # First-party imports
 from gluonts.model.common import Tensor
@@ -80,15 +81,21 @@ class TransformedDistribution(Distribution):
         assert isinstance(self._event_shape, tuple)
         return self._event_shape
 
-    def sample(self, num_samples: Optional[int] = None) -> Tensor:
+    def sample(
+        self, num_samples: Optional[int] = None, dtype=np.float32
+    ) -> Tensor:
         with autograd.pause():
-            s = self.base_distribution.sample(num_samples=num_samples)
+            s = self.base_distribution.sample(
+                num_samples=num_samples, dtype=dtype
+            )
             for t in self.transforms:
                 s = t.f(s)
             return s
 
-    def sample_rep(self, num_samples: Optional[int] = None) -> Tensor:
-        s = self.base_distribution.sample_rep()
+    def sample_rep(
+        self, num_samples: Optional[int] = None, dtype=np.float
+    ) -> Tensor:
+        s = self.base_distribution.sample_rep(dtype=dtype)
         for t in self.transforms:
             s = t.f(s)
         return s
